course 3, lesson 19

+++ main.cpp
@@ -6,7 +6,7 @@ int main()
 {
 
 
-    int var1;
+    unsigned short var1;
     cout<<"Enter your age : ";
     cin>>var1;
     cout<<"After 5 years your age will be "<<var1+5<<" Years old.";
